hud: carve stage.rs (pure move)
The station scene stage: draw_hud, vignette, display thumbnails and
video paint-through, header/control-center/command-deck, compact
surface, orbital scene core, activity lane, focus-detail dispatch.
Sixteen methods leave the StationInner impl into a continuation impl
block; moved methods widen pub(crate).
